[refactor] Combine other enums into Flags.kt and organize everything to match C++ API
