+++ src/main/kotlin/dev/supergrecko/kllvm/ir/ModuleProvider.kt
@@ -1,17 +1,17 @@
-package dev.supergrecko.kllvm.llvm.typedefs
+package dev.supergrecko.kllvm.ir
 
 import dev.supergrecko.kllvm.internal.contracts.Disposable
 import dev.supergrecko.kllvm.internal.contracts.Validatable
-import org.bytedeco.llvm.LLVM.LLVMPassManagerRef
+import org.bytedeco.llvm.LLVM.LLVMModuleProviderRef
 import org.bytedeco.llvm.global.LLVM
 
-public class PassManager internal constructor() :
+public class ModuleProvider internal constructor() :
     AutoCloseable, Validatable, Disposable {
-    internal lateinit var ref: LLVMPassManagerRef
+    internal lateinit var ref: LLVMModuleProviderRef
     public override var valid: Boolean = true
 
-    public constructor(pass: LLVMPassManagerRef) : this() {
-        ref = pass
+    public constructor(provider: LLVMModuleProviderRef) : this() {
+        ref = provider
     }
 
     override fun dispose() {
@@ -19,7 +19,7 @@ public class PassManager internal constructor() :
 
         valid = false
 
-        LLVM.LLVMDisposePassManager(ref)
+        LLVM.LLVMDisposeModuleProvider(ref)
     }
 
     override fun close() = dispose()
