+++ src/main/kotlin/dev/supergrecko/kllvm/ir/PassManager.kt
@@ -1,17 +1,17 @@
-package dev.supergrecko.kllvm.llvm.typedefs
+package dev.supergrecko.kllvm.ir
 
 import dev.supergrecko.kllvm.internal.contracts.Disposable
 import dev.supergrecko.kllvm.internal.contracts.Validatable
-import org.bytedeco.llvm.LLVM.LLVMModuleProviderRef
+import org.bytedeco.llvm.LLVM.LLVMPassManagerRef
 import org.bytedeco.llvm.global.LLVM
 
-public class ModuleProvider internal constructor() :
+public class PassManager internal constructor() :
     AutoCloseable, Validatable, Disposable {
-    internal lateinit var ref: LLVMModuleProviderRef
+    internal lateinit var ref: LLVMPassManagerRef
     public override var valid: Boolean = true
 
-    public constructor(provider: LLVMModuleProviderRef) : this() {
-        ref = provider
+    public constructor(pass: LLVMPassManagerRef) : this() {
+        ref = pass
     }
 
     override fun dispose() {
@@ -19,7 +19,7 @@ public class ModuleProvider internal constructor() :
 
         valid = false
 
-        LLVM.LLVMDisposeModuleProvider(ref)
+        LLVM.LLVMDisposePassManager(ref)
     }
 
     override fun close() = dispose()
