+++ src/main/kotlin/dev/supergrecko/kllvm/ir/types/MetadataType.kt
@@ -2,7 +2,7 @@ package dev.supergrecko.kllvm.ir.types
 
 import dev.supergrecko.kllvm.ir.Type
 import dev.supergrecko.kllvm.ir.TypeKind
-import dev.supergrecko.kllvm.llvm.typedefs.Context
+import dev.supergrecko.kllvm.ir.Context
 import org.bytedeco.llvm.LLVM.LLVMTypeRef
 import org.bytedeco.llvm.global.LLVM
