--- src/main/kotlin/dev/supergrecko/kllvm/llvm/enumerations/AttributeIndex.kt
@@ -1,14 +1,0 @@
-package dev.supergrecko.kllvm.llvm.enumerations
-
-import dev.supergrecko.kllvm.internal.contracts.OrderedEnum
-import org.bytedeco.llvm.global.LLVM
-
-/**
- * Support types matching LLVMAttributeIndex
- *
- * [Documentation](https://llvm.org/doxygen/group__LLVMCCoreTypes.html)
- */
-public enum class AttributeIndex(public override val value: Long) : OrderedEnum<Long> {
-    Return(LLVM.LLVMAttributeReturnIndex),
-    Function(LLVM.LLVMAttributeFunctionIndex)
-}
