--- src/main/kotlin/dev/supergrecko/kllvm/llvm/enumerations/DLLStorageClass.kt
@@ -1,15 +1,0 @@
-package dev.supergrecko.kllvm.llvm.enumerations
-
-import dev.supergrecko.kllvm.internal.contracts.OrderedEnum
-import org.bytedeco.llvm.global.LLVM
-
-/**
- * Support types matching LLVMDLLStorageClass
- *
- * [Documentation](https://llvm.org/doxygen/group__LLVMCCoreTypes.html)
- */
-public enum class DLLStorageClass(public override val value: Int) : OrderedEnum<Int> {
-    Default(LLVM.LLVMDefaultStorageClass),
-    DLLImport(LLVM.LLVMDLLImportStorageClass),
-    DLLExport(LLVM.LLVMDLLExportStorageClass)
-}
