--- src/main/kotlin/dev/supergrecko/kllvm/llvm/enumerations/UnnamedAddr.kt
@@ -1,15 +1,0 @@
-package dev.supergrecko.kllvm.llvm.enumerations
-
-import dev.supergrecko.kllvm.internal.contracts.OrderedEnum
-import org.bytedeco.llvm.global.LLVM
-
-/**
- * Support types matching LLVMUnnamedAddr
- *
- * [Documentation](https://llvm.org/doxygen/group__LLVMCCoreTypes.html)
- */
-public enum class UnnamedAddr(public override val value: Int) : OrderedEnum<Int> {
-    None(LLVM.LLVMNoUnnamedAddr),
-    Local(LLVM.LLVMLocalUnnamedAddr),
-    Global(LLVM.LLVMGlobalUnnamedAddr)
-}
