+++ src/main/kotlin/dev/supergrecko/kllvm/support/Message.kt
@@ -7,9 +7,12 @@ import org.bytedeco.llvm.global.LLVM
 /**
  * A glorified char*
  *
- * A [Message] is a wrapper around Java's ByteBuffer and Bytedeco's BytePointers. It contains a char buffer which will
- * be decoded into a Kotlin [String]. Because these are obtained from C++ they need to be de-allocated which is why
- * they provide the [dispose] method. Failing to call this method will leak memory.
+ * A [Message] is a wrapper around Java's ByteBuffer and Bytedeco's BytePointers
+ * It contains a char buffer which will be decoded into a Kotlin [String].
+ *
+ * Because these are obtained from C++ they need to be de-allocated which is why
+ * they provide the [dispose] method. Failing to call this method will leak
+ * memory.
  *
  * TODO: Test whether this works with ByteBuffers which are not coming from LLVM
  */
@@ -20,8 +23,8 @@ public class Message(private val buffer: ByteBuffer) : Disposable,
     /**
      * Get a string representation of this [Message]
      *
-     * For this method to be available the [Message] needs to be [valid]. If this was called on a deallocated object,
-     * the JVM would crash.
+     * For this method to be available the [Message] needs to be [valid].
+     * If this was called on a de-allocated object, the JVM would crash.
      */
     public fun getString(): String {
         require(valid)
