+++ src/test/kotlin/dev/supergrecko/kllvm/core/ContextTest.kt
@@ -1,7 +1,7 @@
 package dev.supergrecko.kllvm.core
 
 import dev.supergrecko.kllvm.internal.util.runAll
-import dev.supergrecko.kllvm.llvm.typedefs.Context
+import dev.supergrecko.kllvm.ir.Context
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
 import org.junit.jupiter.api.Test
